docs: Update Headlines section description text
Changed from:
'Share critical information about customers and employees that the team needs to know.'

To:
'Headlines should be brief updates and archived upon conclusion of the meeting. If they require discussion they should be added to the Issues list for prioritization.'

Applied to both 90-minute and Express meeting variants for feature parity.